renamed schema folder to schemas, and modified connection.js mongodb port from localhost to 127...
