+++ client/src/index.js
@@ -1,15 +1,27 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
+import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
- const root = ReactDOM.createRoot(document.getElementById('root'));
- root.render(
-   <React.StrictMode>
+
+ReactDOM.render(
+  <React.StrictMode>
     <App />
-  </React.StrictMode>
- );
+  </React.StrictMode>,
+  document.getElementById('root')
+);
+
+
+
+
+
+//  const root = ReactDOM.createRoot(document.getElementById('root'));
+//  root.render(
+//    <React.StrictMode>
+//     <App />
+//   </React.StrictMode>
+//  );
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
