+++ client/src/utils/queries.js
@@ -21,5 +21,5 @@ export const QUERY_COUNTRIES = gql`
   }
 `
 
-export const QUERY_CONTINENTS = gql`
-`
+// export const QUERY_CONTINENTS = gql`
+// `
