+++ server/config/connection.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 
-mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost/Global-Tech-Nomad', {
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/Global-Tech-Nomad', {
   useNewUrlParser: true,
   useUnifiedTopology: true,
   useCreateIndex: true,
