+++ server/models/index.js
@@ -1,4 +1,5 @@
 const User = require('./User');
 const Job = require('./Job');
 const Continents = require('./Continents');
+
 module.exports = { User, Job, Continents};
